Script to extract libargp from Gnulib
This script is inspired by Cygwin's steps for extracting the libargp
together with the build environment from Gnulib and then extracting the
LGPL separately to copy them back into the build environment.

Note that it also seems that the output is influenced by the host
system's autoconf, automake and libtool version. Those will be noted
when the output is committed.