Detect strchrnul() using gl_CHECK_FUNCS_ANDROID_MACOS
This fixes a bug caused by the Github Actions runner for MacOS 15 using
the SDK for MacOS 15.6 but setting a deployment target for MacOS 15.0.0.
This in turn causes the native strchrnul() to be found in the SDK header
but the native function implementation only exists as of MacOS 15.4.

Changes:
* Update strchrnul.m4 to use gl_CHECK_FUNCS_ANDROID_MACOS
* Run `autoreconf --force` to regenerate the build system

Host system:
* autoconf-2.72
* automake-1.18.1
* libtool-2.5.4