Extract libargp from Gnulib at d5c86249fffa98513fd74a424c67c49c3d1bb221
Output directory `libargp-v0.1-6473-gd5c86249ff-20230723T152628-lgpl`
was copied over the `src` directory after the latter was staged as
removed. This allowed git to reflect the actual changes from the old
version to the new.

The intention of extracting libargp at this specific revision is to test
the difference in output between the previous script and the updated
script once the output is merged into the master branch, because some of
that branch's changes are part of the update script as well.

Host system:
    * autoconf-2.72
    * automake-1.18.1
    * libtool-2.5.4