Test C99 format specifiers for argp_error()
This is to confirm whether the compiler allows C99 format specifiers for
libargp functions using the macros in argp-config.h. It should be fine
for GCC and Clang, and this test is mostly aimed at MinGW due to its
dependence on different, sometimes very old, Windows runtimes.

This change also uses CTest directly and adds the --output-on-failure
option.